+++ cvat/apps/dataset_manager/formats/registry.py
@@ -121,6 +121,7 @@ import cvat.apps.dataset_manager.formats.market1501
 import cvat.apps.dataset_manager.formats.icdar
 import cvat.apps.dataset_manager.formats.velodynepoint
 import cvat.apps.dataset_manager.formats.pointcloud
+import cvat.apps.dataset_manager.formats.kitti
 import cvat.apps.dataset_manager.formats.lfw
 import cvat.apps.dataset_manager.formats.cityscapes
 import cvat.apps.dataset_manager.formats.openimages