+++ cvat/apps/dataset_manager/tests/test_formats.py
@@ -296,6 +296,7 @@ class TaskExportTest(_DbTestBase):
             'ICDAR Segmentation 1.0',
             'Kitti Raw Format 1.0',
             'Sly Point Cloud Format 1.0',
+            'KITTI 1.0',
             'LFW 1.0',
             'Cityscapes 1.0',
             'Open Images V6 1.0'
@@ -325,6 +326,7 @@ class TaskExportTest(_DbTestBase):
             'ICDAR Segmentation 1.0',
             'Kitti Raw Format 1.0',
             'Sly Point Cloud Format 1.0',
+            'KITTI 1.0',
             'LFW 1.0',
             'Cityscapes 1.0',
             'Open Images V6 1.0',
@@ -376,6 +378,7 @@ class TaskExportTest(_DbTestBase):
             ('ICDAR Recognition 1.0', 'icdar_word_recognition'),
             ('ICDAR Localization 1.0', 'icdar_text_localization'),
             ('ICDAR Segmentation 1.0', 'icdar_text_segmentation'),
+            # ('KITTI 1.0', 'kitti') format does not support empty annotations
             ('LFW 1.0', 'lfw'),
             # ('Cityscapes 1.0', 'cityscapes'), does not support, empty annotations
         ]:
@@ -785,17 +788,6 @@ class TaskAnnotationsImportTest(_DbTestBase):
                 "type": "rectangle",
                 "occluded": False,
             }]
-        elif annotation_format == "VGGFace2 1.0":
-            shapes = [{
-                "frame": 1,
-                "label_id": task["labels"][1]["id"],
-                "group": None,
-                "source": "manual",
-                "attributes": [],
-                "points": [2.0, 2.1, 40, 50.7],
-                "type": "rectangle",
-                "occluded": False
-            }]
         else:
             rectangle_shape_wo_attrs = {
                 "frame": 1,
@@ -803,7 +795,7 @@ class TaskAnnotationsImportTest(_DbTestBase):
                 "group": 0,
                 "source": "manual",
                 "attributes": [],
-                "points": [2.0, 2.1, 40, 50.7],
+                "points": [2.0, 2.1, 40, 10.7],
                 "type": "rectangle",
                 "occluded": False,
             }
@@ -823,7 +815,7 @@ class TaskAnnotationsImportTest(_DbTestBase):
                         "value": task["labels"][0]["attributes"][1]["default_value"]
                     }
                 ],
-                "points": [1.0, 2.1, 10.6, 53.22],
+                "points": [1.0, 2.1, 10.6, 13.22],
                 "type": "rectangle",
                 "occluded": False,
             }
@@ -838,7 +830,7 @@ class TaskAnnotationsImportTest(_DbTestBase):
                     {
                         "frame": 0,
                         "attributes": [],
-                        "points": [1.0, 2.1, 10.6, 53.22, 90, 90.222],
+                        "points": [1.0, 2.1, 10.6, 53.22, 30, 20.222],
                         "type": "polygon",
                         "occluded": False,
                         "outside": False
@@ -871,7 +863,7 @@ class TaskAnnotationsImportTest(_DbTestBase):
             }
 
             if annotation_format == "VGGFace2 1.0":
-                shapes = rectangle_shape_wo_attrs
+                shapes = [rectangle_shape_wo_attrs]
             elif annotation_format == "CVAT 1.1":
                 shapes = [rectangle_shape_wo_attrs,
                     rectangle_shape_with_attrs]