+++ cvat/apps/engine/tests/test_rest_api.py
@@ -4822,6 +4822,10 @@ class TaskAnnotationAPITestCase(JobAnnotationAPITestCase):
                 annotations["shapes"] = points_wo_attrs \
                                       + tags_wo_attrs
 
+            elif annotation_format == "KITTI 1.0":
+                annotations["shapes"] = rectangle_shapes_wo_attrs \
+                                            + polygon_shapes_wo_attrs
+
             elif annotation_format == "Market-1501 1.0":
                 tags_with_attrs = [{
                     "frame": 1,